remove unnecessary keyboard listener and refactor files
