+++ src/main/frontend/components/employeeSettings/EditEmailModal.js
@@ -8,12 +8,11 @@ import {
     View,
     Platform,
     KeyboardAvoidingView,
-    Keyboard, TouchableWithoutFeedback
 } from "react-native";
-import {black, destructiveAction, primaryGreen, secondaryGray} from "../utils/Colors";
+import {black, destructiveAction, primaryGreen, secondaryGray} from "../../utils/Colors";
 import {FontAwesomeIcon} from "@fortawesome/react-native-fontawesome";
-import {Check, XMark} from "../utils/Icons";
-import employeeData from "../mockApiCalls/employeeData.json";
+import {Check, XMark} from "../../utils/Icons";
+import employeeData from "../../mockApiCalls/employeeData.json";
 
 function EditEmailModal({emailModalVisible, setEmailModalVisible}) {
     const [email, onChangeEmail] = React.useState(employeeData.email);
@@ -29,33 +28,31 @@ function EditEmailModal({emailModalVisible, setEmailModalVisible}) {
             <KeyboardAvoidingView
                 behavior={Platform.OS === 'ios' ? 'padding' : 'height'}
                 style={styles.container}>
-                <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
-                    <View style={styles.centeredView}>
-                        <View style={styles.modalView}>
-                            <Text style={styles.modalText}>Edit Email</Text>
-                            <TextInput
-                                style={styles.inputText}
-                                autoCapitalize={"words"}
-                                onChangeText={onChangeEmail}
-                                value={email}
-                                placeholder="ex. johndoe@email.com"
-                                placeholderTextColor={secondaryGray}
-                            />
-                            <View style={styles.buttonsContainer}>
-                                <TouchableOpacity
-                                    style={styles.buttonCancel}
-                                    onPress={() => setEmailModalVisible(!emailModalVisible)}>
-                                    <FontAwesomeIcon icon={XMark} size={32} color={destructiveAction} />
-                                </TouchableOpacity>
-                                <TouchableOpacity
-                                    style={styles.buttonSave}
-                                    onPress={() => setEmailModalVisible(!emailModalVisible)}>
-                                    <FontAwesomeIcon icon={Check} size={32} color={primaryGreen} />
-                                </TouchableOpacity>
-                            </View>
+                <View style={styles.centeredView}>
+                    <View style={styles.modalView}>
+                        <Text style={styles.modalText}>Edit Email</Text>
+                        <TextInput
+                            style={styles.inputText}
+                            autoCapitalize={"words"}
+                            onChangeText={onChangeEmail}
+                            value={email}
+                            placeholder="ex. johndoe@email.com"
+                            placeholderTextColor={secondaryGray}
+                        />
+                        <View style={styles.buttonsContainer}>
+                            <TouchableOpacity
+                                style={styles.buttonCancel}
+                                onPress={() => setEmailModalVisible(!emailModalVisible)}>
+                                <FontAwesomeIcon icon={XMark} size={32} color={destructiveAction} />
+                            </TouchableOpacity>
+                            <TouchableOpacity
+                                style={styles.buttonSave}
+                                onPress={() => setEmailModalVisible(!emailModalVisible)}>
+                                <FontAwesomeIcon icon={Check} size={32} color={primaryGreen} />
+                            </TouchableOpacity>
                         </View>
                     </View>
-                </TouchableWithoutFeedback>
+                </View>
             </KeyboardAvoidingView>
         </Modal>
     )
